+++ ci4/app/Entities/DeploymentPackage.php
@@ -17,6 +17,7 @@ use RestExtension\Core\Entity;
  * @property Workspace $workspaces
  * @property DeploymentPackageDeploymentSpecification $deployment_package_deployment_specifications
  * @property DeploymentPackageEnvironmentVariable $deployment_package_environment_variables
+ * @property Label $labels
  */
 class DeploymentPackage extends Entity {
 
@@ -30,6 +31,12 @@ class DeploymentPackage extends Entity {
         $this->deployment_package_environment_variables->find()->deleteAll();
         $this->save($values);
         $this->deployment_package_environment_variables = $values;
+    }
+
+    public function updateLabels(Label $values): void {
+        $this->labels->find()->deleteAll();
+        $this->save($values);
+        $this->labels = $values;
     }
 
     /**