+++ ci4/app/Models/LabelModel.php
@@ -11,6 +11,7 @@ class LabelModel extends Model implements ResourceModelInterface {
 
     public $hasMany = [
         WorkspaceModel::class,
+        DeploymentPackageModel::class,
     ];
 
     public function preRestGet($queryParser, $id) {