+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v0.1.13  ()
+
+### Fixed bugs
+* 
+
+### Enhancements
+* Add labels to workspace templates [Issue #15](https://github.com/4spacesdk/kubernetes-service-orchestrator/issues/15)
+
+### Upgrade guide
+1. Deploy new image
+2. Migrate database. [(Guide)](https://github.com/4spacesdk/kubernetes-service-orchestrator?tab=readme-ov-file#migrate-database-helm)
+
+
+
 ## v0.1.12  (2024-02-21)
 
 ### Fixed bugs
@@ -8,7 +22,7 @@
 ### Enhancements
 * Moved kubernetes.io/ingress.class annotation to spec ingressClassName [link](https://kubernetes.io/docs/concepts/services-networking/ingress/#deprecated-annotation).
 * You can now specify `imagePullSecret` for container images
-  * Set default value with `IMAGE_PULL_SECRET_DEFAULT_NAME` environment variable or set value in helm `values.yaml` 
+    * Set default value with `IMAGE_PULL_SECRET_DEFAULT_NAME` environment variable or set value in helm `values.yaml`
 
 ### Upgrade guide
 1. Deploy new image